moving all exammples as stub data related tests to ExamplesStubStubTest
