+++ core/src/test/kotlin/in/specmatic/stub/HttpStubTest.kt
@@ -847,168 +847,6 @@ paths:
 """.trim(), ""
         ).toFeature()
 
-        private val featureWithQueryParamExamples = OpenApiSpecification.fromYAML(
-            """
-openapi: 3.0.1
-info:
-  title: Data API
-  version: "1"
-paths:
-  /:
-    get:
-      summary: Data
-      parameters:
-        - name: type
-          schema:
-            type: string
-          in: query
-          required: true
-          examples:
-            QUERY_SUCCESS:
-              value: data
-      responses:
-        "200":
-          description: Data
-          content:
-            application/json:
-              schema:
-                type: array
-                items:
-                  type:
-                    string
-              examples:
-                QUERY_SUCCESS:
-                  value: ["one", "two"]
-""".trim(), ""
-        ).toFeature()
-
-        private val featureWithPathParamExamples = OpenApiSpecification.fromYAML(
-            """
-openapi: 3.0.1
-info:
-  title: Data API
-  version: "1"
-paths:
-  /{productId}:
-    get:
-      summary: Data
-      parameters:
-        - name: productId
-          schema:
-            type: string
-          in: path
-          required: true
-          examples:
-            PATH_SUCCESS:
-              value: xyz123
-            PATH_FAILURE:
-              value: abc123
-      responses:
-        "200":
-          description: Data
-          content:
-            application/json:
-              schema:
-                type: array
-                items:
-                  type:
-                    string
-              examples:
-                PATH_SUCCESS:
-                  value: ["one", "two"]
-        "404":
-          description: Data not found
-          content:
-            text/plain:
-              schema:
-                type: string
-              examples:
-                PATH_FAILURE:
-                  value: "Could not find abc123"
-""".trim(), ""
-        ).toFeature()
-
-        private val featureWithHeaderParamExamples = OpenApiSpecification.fromYAML(
-            """
-openapi: 3.0.1
-info:
-  title: Header Example API
-  version: "1"
-paths:
-  /hello:
-    get:
-      parameters:
-        - name: userId
-          schema:
-            type: string
-          in: header
-          required: true
-          examples:
-            HEADER_SUCCESS:
-              value: John
-            HEADER_FAILURE:
-              value: Jane
-      responses:
-        "200":
-          description: Data
-          content:
-            text/plain:
-              schema:
-                type: string
-              examples:
-                HEADER_SUCCESS:
-                  value: "Hello John"
-        "401":
-          description: Unauthorized
-          content:
-            text/plain:
-              schema:
-                type: string
-              examples:
-                HEADER_FAILURE:
-                  value: "User Jane not authorized"
-""".trim(), ""
-        ).toFeature()
-
-        private val featureWithOmittedParamExamples = OpenApiSpecification.fromYAML(
-            """
-openapi: 3.0.1
-info:
-  title: Header Example API
-  version: "1"
-paths:
-  /hello:
-    get:
-      parameters:
-        - name: userId
-          schema:
-            type: string
-          in: header
-          required: true
-          examples:
-            HEADER_SUCCESS:
-              value: John
-        - name: role
-          schema:
-            type: string
-          in: header
-          required: false
-          examples:
-            HEADER_SUCCESS:
-              value: "(omit)"
-      responses:
-        "200":
-          description: Data
-          content:
-            text/plain:
-              schema:
-                type: string
-              examples:
-                HEADER_SUCCESS:
-                  value: "Hello John"
-""".trim(), ""
-        ).toFeature()
-
         @Test
         fun `expectations for payload from examples`() {
             HttpStub(featureWithBodyExamples).use { stub ->
@@ -1016,60 +854,6 @@ paths:
                     .let { response ->
                         assertThat(response.status).isEqualTo(200)
                         assertThat(response.body).isEqualTo(parsedJSONObject("""{"message":"example_expectation"}"""))
-                    }
-            }
-        }
-
-        @Test
-        fun `expectations for query params from examples`() {
-            HttpStub(featureWithQueryParamExamples).use { stub ->
-                stub.client.execute(HttpRequest("GET", "/?type=data"))
-                    .let { response ->
-                        assertThat(response.status).isEqualTo(200)
-                        assertThat(response.body).isEqualTo(parsedJSONArray("""["one", "two"]"""))
-                    }
-            }
-        }
-
-        @Test
-        fun `expectations for path params from examples`() {
-            HttpStub(featureWithPathParamExamples).use { stub ->
-                stub.client.execute(HttpRequest("GET", "/xyz123"))
-                    .let { response ->
-                        assertThat(response.status).isEqualTo(200)
-                        assertThat(response.body).isEqualTo(parsedJSONArray("""["one", "two"]"""))
-                    }
-                stub.client.execute(HttpRequest("GET", "/abc123"))
-                    .let { response ->
-                        assertThat(response.status).isEqualTo(404)
-                        assertThat(response.body.toStringLiteral()).isEqualTo("Could not find abc123")
-                    }
-            }
-        }
-
-        @Test
-        fun `expectations for header params from examples`() {
-            HttpStub(featureWithHeaderParamExamples).use { stub ->
-                stub.client.execute(HttpRequest("GET", "/hello", headers = mapOf("userId" to "John")))
-                    .let { response ->
-                        assertThat(response.status).isEqualTo(200)
-                        assertThat(response.body.toStringLiteral()).isEqualTo("Hello John")
-                    }
-                stub.client.execute(HttpRequest("GET", "/hello", headers = mapOf("userId" to "Jane")))
-                    .let { response ->
-                        assertThat(response.status).isEqualTo(401)
-                        assertThat(response.body.toStringLiteral()).isEqualTo("User Jane not authorized")
-                    }
-            }
-        }
-
-        @Test
-        fun `expectations from examples ignores omitted parameters while matching stub request`() {
-            HttpStub(featureWithOmittedParamExamples).use { stub ->
-                stub.client.execute(HttpRequest("GET", "/hello", headers = mapOf("userId" to "John")))
-                    .let { response ->
-                        assertThat(response.status).isEqualTo(200)
-                        assertThat(response.body.toStringLiteral()).isEqualTo("Hello John")
                     }
             }
         }
